fixed granting/revoking to only one user/tag

+++ transactions/revokeMinter.cdc
@@ -3,16 +3,19 @@ import "BasicNFT"
 import "UniversalCollection"
 import "MetadataViews"
 
-transaction() {
+transaction(name: String) {
 
     prepare(signer: auth(Capabilities) &Account) {
         //we get all capabilities for a given path and we delete them
         //
         let storagePath = /storage/basicNFTMinter
-        let capability = signer.capabilities.storage.forEachController(forPath:storagePath, 
-        fun(scc: &StorageCapabilityController) : Bool {
-            scc.delete()
-            return false
+        let capability = signer.capabilities.storage.forEachController(forPath:storagePath, fun(scc: &StorageCapabilityController) : Bool {
+
+            if scc.tag == name {
+                scc.delete()
+                return false
+            }
+            return true
         })
     }
 }
